requirements.txt
Flask
numpy
Pillow
tensorflow==2.13.0
